Moving molecule building into builder.py + refactoring
